+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Split out from [pureharm](https://github.com/busymachines/pureharm) as of version `0.0.7`.
 
+Newly cross published for both Scala 2.13, and 3.0.0-RC1 on JVM and JS runtimes.
+
 `busymachines.pureharm.anomaly`:
 - add `AnomalyLike` super-type that is inherited by `Anomaly`, `Catastrophe`, `Anomalies`, this allows
   for easy non-anomaly catch-all that allows rethrowing.